requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
pdfplumber
spacy
reportlab
whisper
